requirements.txt
python-telegram-bot==20.3
python-docx
openai
python-dotenv
